integer player

+++ app_player_adapter.hpp
@@ -40,7 +40,7 @@ private:
 	kb::ModPlayer& player_;
 	alg::RingIndex<4096> rw_;
 	std::int16_t buf_[4096*2];
-	float pbuf_[4096*2]; };
+	int pbuf_[4096*2]; };
 
 
 }  // namespace app
